OneDrive/Desktop/Medflux/requirements.txt
∩╗┐fastapi==0.116.1
httptools==0.6.4
watchfiles==1.1.0
websockets==15.0.1
PyYAML==6.0.2
python-dotenv==1.1.1
uvicorn==0.35.0
python-multipart==0.0.20
python-docx==1.2.0
pymupdf
Pillow
chardet

